requirements.txt
discord.py
flask
pyyaml